+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - none
 
 # v2025.5.1-beta
+- ci - Add files *.dat.
+- av - Rename mykxx.
 - av - Replace "Cheat Engine" in .exe.
 - av - Rename mydfx3drv.
 - av - Rename CEDRIVER73, dbk32, dbk64, ManufacturerName.