+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 # Todo
 - none
 
-# v2025.5.1-beta
+# v2025.5.31
+- ci - Upgrade to cheat-engine v2024.8.25.
 - ci - sh jslint_ci.sh shCheatengineUpdate /g/games/cheatengine/.
 - av - Rename 1337.
 - ci - Add files *.dat.
@@ -13,6 +14,6 @@
 - av - Rename CEDRIVER73, dbk32, dbk64, ManufacturerName.
 - ci - Upgrade ci to nodejs v22.
 
-# v2023.11.1-beta
+# v2023.11.9
 - ce - Add shell-function shCheatengineUpdate().
 - ci - Create github-action to autobuild cheat-engine.